requirements.txt
asgiref==3.2.3
pytz==2019.3
sqlparse==0.3.0
django-bootstrap4==2.0.1
django-bootstrap-datepicker-plus==3.0.5


